+++ psm_utils/__init__.py
@@ -1,6 +1,16 @@
 """Common utilities for parsing and handling PSMs, and search engine results."""
 
-__version__ = "0.2.0"
+__version__ = "0.4.1"
+
+from warnings import filterwarnings
+
+# mzmlb is not used, so hdf5plugin is not needed
+filterwarnings(
+    "ignore",
+    message="hdf5plugin is missing",
+    category=UserWarning,
+    module="psims.mzmlb",
+)
 
 from psm_utils.peptidoform import Peptidoform
 from psm_utils.psm import PSM