+++ psm_utils/peptidoform.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
+import numpy as np
 from pyteomics import mass, proforma
 
 from psm_utils.exceptions import PSMUtilsException
+from psm_utils.utils import mass_to_mz
 
 
 class Peptidoform:
@@ -33,7 +35,12 @@ class Peptidoform:
             Dict with sequence-wide properties.
 
         """
-        self.parsed_sequence, self.properties = proforma.parse(proforma_sequence)
+        try:
+            self.parsed_sequence, self.properties = proforma.parse(proforma_sequence)
+        except proforma.ProFormaError as e:
+            raise PeptidoformException(
+                f"Could not parse ProForma sequence: {proforma_sequence}"
+            ) from e
 
         if self.properties["isotopes"]:
             raise NotImplementedError(
@@ -53,7 +60,7 @@ class Peptidoform:
         try:
             return self.proforma == __o.proforma
         except AttributeError:
-            raise NotImplemented("Object is not a Peptidoform")
+            raise NotImplementedError("Object is not a Peptidoform")
 
     @property
     def proforma(self) -> str:
@@ -101,6 +108,26 @@ class Peptidoform:
             return None
 
     @property
+    def is_modified(self) -> bool:
+        """
+        Whether or not the peptidoform carries any modification of any type.
+
+        Includes N- and C-terminal, fixed, sequential, labile, and unlocalized
+        modifications.
+
+        """
+        mod_properties = [
+            "n_term",
+            "c_term",
+            "unlocalized_modifications",
+            "labile_modifications",
+            "fixed_modifications",
+        ]
+        has_sequential = any(mods for _, mods in self.parsed_sequence)
+        has_other = any([self.properties[prop] for prop in mod_properties])
+        return has_sequential or has_other
+
+    @property
     def sequential_composition(self) -> list[mass.Composition]:
         """
         Atomic compositions of both termini and each residue, including modifications.
@@ -134,10 +161,10 @@ class Peptidoform:
             for tag in self.properties["n_term"]:
                 try:
                     n_term += tag.composition
-                except (AttributeError, KeyError):
+                except (AttributeError, KeyError) as e:
                     raise ModificationException(
                         f"Cannot resolve composition for modification {tag.value}."
-                    )
+                    ) from e
         comp_list.append(n_term)
 
         # Sequence
@@ -145,10 +172,10 @@ class Peptidoform:
             # Amino acid
             try:
                 position_comp = mass.std_aa_comp[aa].copy()
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise AmbiguousResidueException(
                     f"Cannot resolve composition for amino acid {aa}."
-                )
+                ) from e
             # Fixed modifications
             if aa in fixed_rules:
                 position_comp += fixed_rules[aa]
@@ -157,11 +184,11 @@ class Peptidoform:
                 for tag in tags:
                     try:
                         position_comp += tag.composition
-                    except (AttributeError, KeyError):
+                    except (AttributeError, KeyError) as e:
                         raise ModificationException(
                             "Cannot resolve composition for modification "
                             f"{tag.value}."
-                        )
+                        ) from e
             comp_list.append(position_comp)
 
         # C-terminus
@@ -170,10 +197,10 @@ class Peptidoform:
             for tag in self.properties["c_term"]:
                 try:
                     c_term += tag.composition
-                except (AttributeError, KeyError):
+                except (AttributeError, KeyError) as e:
                     raise ModificationException(
                         f"Cannot resolve composition for modification {tag.value}."
-                    )
+                    ) from e
         comp_list.append(c_term)
 
         return comp_list
@@ -197,17 +224,17 @@ class Peptidoform:
         for tag in self.properties["labile_modifications"]:
             try:
                 comp += tag.composition
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise ModificationException(
                     f"Cannot resolve composition for modification {tag.value}."
-                )
+                ) from e
         for tag in self.properties["unlocalized_modifications"]:
             try:
                 comp += tag.composition
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise ModificationException(
                     f"Cannot resolve composition for modification {tag.value}."
-                )
+                ) from e
         return comp
 
     @property
@@ -243,10 +270,10 @@ class Peptidoform:
             for tag in self.properties["n_term"]:
                 try:
                     n_term += tag.mass
-                except (AttributeError, KeyError):
+                except (AttributeError, KeyError) as e:
                     raise ModificationException(
                         f"Cannot resolve mass for modification {tag.value}."
-                    )
+                    ) from e
         mass_list.append(n_term)
 
         # Sequence
@@ -254,10 +281,10 @@ class Peptidoform:
             # Amino acid
             try:
                 position_mass = mass.std_aa_mass[aa]
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise AmbiguousResidueException(
                     f"Cannot resolve mass for amino acid {aa}."
-                )
+                ) from e
             # Fixed modifications
             if aa in fixed_rules:
                 position_mass += fixed_rules[aa]
@@ -266,10 +293,10 @@ class Peptidoform:
                 for tag in tags:
                     try:
                         position_mass += tag.mass
-                    except (AttributeError, KeyError):
+                    except (AttributeError, KeyError) as e:
                         raise ModificationException(
                             "Cannot resolve mass for modification " f"{tag.value}."
-                        )
+                        ) from e
             mass_list.append(position_mass)
 
         # C-terminus
@@ -278,10 +305,10 @@ class Peptidoform:
             for tag in self.properties["c_term"]:
                 try:
                     c_term += tag.mass
-                except (AttributeError, KeyError):
+                except (AttributeError, KeyError) as e:
                     raise ModificationException(
                         f"Cannot resolve mass for modification {tag.value}."
-                    )
+                    ) from e
         mass_list.append(c_term)
 
         return mass_list
@@ -302,17 +329,17 @@ class Peptidoform:
         for tag in self.properties["labile_modifications"]:
             try:
                 mass += tag.mass
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise ModificationException(
                     f"Cannot resolve mass for modification {tag.value}."
-                )
+                ) from e
         for tag in self.properties["unlocalized_modifications"]:
             try:
                 mass += tag.mass
-            except (AttributeError, KeyError):
+            except (AttributeError, KeyError) as e:
                 raise ModificationException(
                     f"Cannot resolve mass for modification {tag.value}."
-                )
+                ) from e
         return mass
 
     @property
@@ -332,10 +359,7 @@ class Peptidoform:
 
         """
         if self.precursor_charge:
-            return (
-                self.theoretical_mass
-                + (mass.nist_mass["H"][1][0] * self.precursor_charge)
-            ) / self.precursor_charge
+            return mass_to_mz(self.theoretical_mass, self.precursor_charge)
         else:
             return None
 
@@ -370,8 +394,12 @@ class Peptidoform:
             new_mods = []
             for mod in mods:
                 try:
-                    if mod.value in mapping:
-                        new_mods.append(proforma.process_tag_tokens(mapping[mod.value]))
+                    if isinstance(mod, proforma.MassModification):
+                        mod_value = _format_number_as_string(mod.value)
+                    else:
+                        mod_value = mod.value
+                    if mod_value in mapping:
+                        new_mods.append(proforma.process_tag_tokens(mapping[mod_value]))
                     else:
                         new_mods.append(mod)
                 except AttributeError:
@@ -477,6 +505,13 @@ class Peptidoform:
 
             # Remove fixed modifications
             self.properties["fixed_modifications"] = []
+
+
+def _format_number_as_string(num):
+    """Format number as string for ProForma mass modifications."""
+    sign = "+" if np.sign(num) == 1 else "-"
+    num = str(num).rstrip("0").rstrip(".")
+    return sign + num
 
 
 class PeptidoformException(PSMUtilsException):