+++ psm_utils/psm_list.py
@@ -81,6 +81,9 @@ class PSMList(BaseModel):
     def __str__(self):
         return self.__repr__()
 
+    def __add__(self, other):
+        return PSMList(psm_list=self.psm_list + other.psm_list)
+
     def __iter__(self) -> Iterable[PSM]:
         return self.psm_list.__iter__()
 
@@ -88,7 +91,7 @@ class PSMList(BaseModel):
         return self.psm_list.__len__()
 
     def __getitem__(self, item) -> PSM | list[PSM]:
-        if isinstance(item, int):
+        if isinstance(item, (int, np.integer)):
             # Return single PSM by index
             return self.psm_list[item]
         elif isinstance(item, slice):
@@ -218,7 +221,7 @@ class PSMList(BaseModel):
 
         """
         for key in ["score", "is_decoy"]:
-            if (self[key] == None).any():
+            if (self[key] == None).any():  # noqa: E711 (self[key] is a Numpy array)
                 raise ValueError(
                     f"Cannot calculate q-values if not all PSMs have `{key}` assigned."
                 )